requirements.txt
# ==============================================================
#  Resume Parser AI — Python Dependencies
# ==============================================================
#
#  QUICK INSTALL (run these commands one by one):
#
#  Step 1 — Install packages:
#      pip install -r requirements.txt
#
#  Step 2 — Download the spaCy language model:
#      python -m spacy download en_core_web_sm
#
# ==============================================================

# Web Framework
Flask>=2.3.0
Werkzeug>=2.3.0

# PDF text extraction
PyPDF2>=3.0.0

# Word document (.docx) reading
python-docx>=0.8.11

# NLP — Named Entity Recognition
spacy>=3.7.0

# Utilities (usually pre-installed, listed for completeness)
Jinja2>=3.1.0
itsdangerous>=2.1.0
click>=8.1.0
